+++ processor/src/test/resources/expected/com/example/collections/Arez_MemoizeNonnullListModel.java
@@ -87,14 +87,14 @@ public final class Arez_MemoizeNonnullListModel extends MemoizeNonnullListModel 
     }
     if ( Arez.areCollectionsPropertiesUnmodifiable() ) {
       if ( null == this.$$arezd$$_$$cache$$_myValue ) {
-        this.$$arezd$$_$$cache$$_myValue = CollectionsUtil.wrap( (List<String>) this.$$arez$$_myValue.get() );
+        this.$$arezd$$_$$cache$$_myValue = CollectionsUtil.wrap( this.$$arez$$_myValue.get() );
       } else {
         // Make sure that we are observing computable value
         this.$$arez$$_myValue.get();
       }
       return $$arezd$$_$$cache$$_myValue;
     } else {
-      return (List<String>) this.$$arez$$_myValue.get();
+      return this.$$arez$$_myValue.get();
     }
   }